+++ src/components/trays/remove/index.js
