+++ src/components/trays/remove/remove-observations.js
